+++ src/calendar.js
@@ -1253,7 +1253,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var visibleGroups = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_Groups );
 
                 if ( !areArraysTheSame( _options_Configuration.visibleGroups, visibleGroups ) ) {
-                    _options_Configuration.visibleGroups = visibleGroups;
+                    _options_Configuration.visibleGroups = [].slice.call( visibleGroups );
                     itemWasChanged = true;
 
                     fireCustomTrigger( "onVisibleGroupsChanged", _options_Configuration.visibleGroups );
@@ -1264,7 +1264,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var visibleEventTypes = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_EventTypes, true );
 
                 if ( !areArraysTheSame( _options_Configuration.visibleEventTypes, visibleEventTypes ) ) {
-                    _options_Configuration.visibleEventTypes = visibleEventTypes;
+                    _options_Configuration.visibleEventTypes = [].slice.call( visibleEventTypes );
                     itemWasChanged = true;
 
                     fireCustomTrigger( "onVisibleEventTypesChanged", _options_Configuration.visibleEventTypes );
@@ -1275,7 +1275,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var visibleDays = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_Days, true );
 
                 if ( visibleDays.length >= 1 && !areArraysTheSame( _options.visibleDays, visibleDays ) ) {
-                    _options.visibleDays = visibleDays;
+                    _options.visibleDays = [].slice.call( visibleDays );
                     _element_Calendar_PreviousDaysVisibleBeforeSingleDayView = [];
                     
                     fireOptionsCustomTrigger = true;
@@ -1287,7 +1287,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
                 var workingDays = getSideMenuCheckedCheckBoxNames( _element_SideMenu_Content_Section_WorkingDays, true );
 
                 if ( !areArraysTheSame( _options.workingDays, workingDays ) ) {
-                    _options.workingDays = workingDays;
+                    _options.workingDays = [].slice.call( workingDays );
 
                     fireOptionsCustomTrigger = true;
                     itemWasChanged = true;
@@ -12590,8 +12590,12 @@ function calendarJs( elementOrId, options, searchOptions ) {
                         group = getString( event.group ),
                         url = getString( event.url );
 
-                    if ( isDefined( _options_Configuration.visibleGroups ) ) {
-                        _options_Configuration.visibleGroups.push( getGroupName( group ) );
+                    if ( isDefined( _options_Configuration.visibleGroups ) && !isCustomTriggerSet( "onEventsFetch" ) ) {
+                        var groupName = getGroupName( group );
+
+                        if ( _options_Configuration.visibleGroups.indexOf( groupName ) === -1 ) {
+                            _options_Configuration.visibleGroups.push( groupName );
+                        }
                     }
     
                     if ( !isDefined( event.id ) ) {